refactor: front/(auth) 리팩토링 완료
