+++ frontend/src/app/(auth)/_components/AuthFormLayout/AuthFormLayout.module.css
