+++ frontend/src/app/(auth)/_components/AuthFormLayout/AuthFormLayout.tsx
