+++ frontend/src/app/(auth)/_components/AuthFormLayout/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AuthFormLayout";
