+++ frontend/src/app/(auth)/_components/GoogleLoginButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./GoogleLoginButton";
