+++ frontend/src/app/(auth)/_components/TermsModal/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./TermsModal";
+export type { TermType } from "./termsData";
