+++ frontend/src/app/(auth)/forgot-password/_components/ForgotPasswordForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ForgotPasswordForm";
