+++ frontend/src/app/(auth)/login/_components/LoginForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LoginForm";
