--- frontend/src/app/(auth)/signup/page.module.css
@@ -1,2 +1,0 @@
-/* 현재 페이지에 특화된 개별 CSS가 필요하지 않아 이 파일은 비워져 있습니다. 
-   모든 공통 폼 스타일링은 ../_components/AuthFormLayout.module.css 에서 일괄 처리됩니다. */
