+++ frontend/src/app/(auth)/verify-email/_components/VerifyEmailContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./VerifyEmailContent";
